test: 버킷리스트 완료처리, 수정, 삭제 관련 시나리오 기반 테스트
