+++ src/main/java/com/hanaieum/server/domain/bucketList/service/BucketListServiceImpl.java
@@ -6,35 +6,38 @@ import com.hanaieum.server.domain.account.entity.Account;
 import com.hanaieum.server.domain.account.service.AccountService;
 import com.hanaieum.server.domain.autoTransfer.service.AutoTransferScheduleService;
 import com.hanaieum.server.domain.bucketList.calculator.InterestCalculator;
-import com.hanaieum.server.domain.coupon.service.CouponService;
-import com.hanaieum.server.domain.transaction.entity.Transaction;
-import com.hanaieum.server.domain.transaction.entity.TransactionType;
-import com.hanaieum.server.domain.transaction.service.TransactionService;
-import com.hanaieum.server.domain.transfer.service.TransferService;
 import com.hanaieum.server.domain.bucketList.dto.*;
 import com.hanaieum.server.domain.bucketList.entity.BucketList;
 import com.hanaieum.server.domain.bucketList.entity.BucketListStatus;
 import com.hanaieum.server.domain.bucketList.entity.BucketParticipant;
 import com.hanaieum.server.domain.bucketList.repository.BucketListRepository;
 import com.hanaieum.server.domain.bucketList.repository.BucketParticipantRepository;
+import com.hanaieum.server.domain.coupon.service.CouponService;
 import com.hanaieum.server.domain.group.entity.Group;
 import com.hanaieum.server.domain.member.entity.Member;
 import com.hanaieum.server.domain.member.repository.MemberRepository;
+import com.hanaieum.server.domain.transaction.entity.Transaction;
+import com.hanaieum.server.domain.transaction.entity.TransactionType;
+import com.hanaieum.server.domain.transaction.service.TransactionService;
+import com.hanaieum.server.domain.transfer.service.TransferService;
 import com.hanaieum.server.security.CustomUserDetails;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
-import java.util.List;
 import java.util.ArrayList;
+import java.util.List;
 
-import static java.lang.Integer.*;
+import static com.hanaieum.server.common.exception.ErrorCode.*;
+import static com.hanaieum.server.domain.bucketList.entity.BucketListStatus.COMPLETED;
+import static com.hanaieum.server.domain.bucketList.entity.BucketListStatus.IN_PROGRESS;
+import static java.lang.Integer.parseInt;
+import static org.springframework.security.core.context.SecurityContextHolder.getContext;
 
 @Slf4j
 @Service
@@ -58,12 +61,12 @@ public class BucketListServiceImpl implements BucketListService {
      * 현재 로그인한 사용자 정보를 가져오는 공통 메서드
      */
     private Member getCurrentMember() {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Authentication authentication = getContext().getAuthentication();
         CustomUserDetails userDetails = (CustomUserDetails) authentication.getPrincipal();
         Long memberId = userDetails.getId();
 
         return memberRepository.findById(memberId)
-                .orElseThrow(() -> new CustomException(ErrorCode.MEMBER_NOT_FOUND));
+                .orElseThrow(() -> new CustomException(MEMBER_NOT_FOUND));
     }
 
     /**
@@ -71,7 +74,7 @@ public class BucketListServiceImpl implements BucketListService {
      */
     private void validateBucketListOwnership(BucketList bucketList, Long memberId) {
         if (!bucketList.getMember().getId().equals(memberId)) {
-            throw new CustomException(ErrorCode.BUCKET_LIST_ACCESS_DENIED);
+            throw new CustomException(BUCKET_LIST_ACCESS_DENIED);
         }
     }
 
@@ -80,7 +83,7 @@ public class BucketListServiceImpl implements BucketListService {
      */
     private BucketList getBucketListWithOwnershipValidation(Long bucketListId, Long memberId) {
         BucketList bucketList = bucketListRepository.findByIdAndDeletedWithParticipants(bucketListId, false)
-                .orElseThrow(() -> new CustomException(ErrorCode.BUCKET_LIST_NOT_FOUND));
+                .orElseThrow(() -> new CustomException(BUCKET_LIST_NOT_FOUND));
 
         validateBucketListOwnership(bucketList, memberId);
         return bucketList;
@@ -95,11 +98,11 @@ public class BucketListServiceImpl implements BucketListService {
 
         // 삭제되지 않은 버킷리스트만 조회 (참여자 정보 포함)
         BucketList bucketList = bucketListRepository.findByIdAndDeletedWithParticipants(bucketListId, false)
-                .orElseThrow(() -> new CustomException(ErrorCode.BUCKET_LIST_NOT_FOUND));
+                .orElseThrow(() -> new CustomException(BUCKET_LIST_NOT_FOUND));
 
         // 소유자 확인
         if (!bucketList.getMember().getId().equals(memberId)) {
-            throw new CustomException(ErrorCode.BUCKET_LIST_ACCESS_DENIED);
+            throw new CustomException(BUCKET_LIST_ACCESS_DENIED);
         }
         return bucketList;
     }
@@ -125,7 +128,7 @@ public class BucketListServiceImpl implements BucketListService {
                 .targetDate(targetDate)
                 .publicFlag(requestDto.getPublicFlag())
                 .shareFlag(requestDto.getTogetherFlag())
-                .status(BucketListStatus.IN_PROGRESS) // 생성 시 기본값: 진행중
+                .status(IN_PROGRESS) // 생성 시 기본값: 진행중
                 .deleted(false) // 기본값: 활성화
                 .build();
 
@@ -144,7 +147,7 @@ public class BucketListServiceImpl implements BucketListService {
                         savedBucketList.getId(), e.getMessage());
                 // 참여자 추가 실패해도 버킷리스트 생성은 성공 처리
             }
-            }
+        }
 
         // 머니박스 자동 생성
         if (requestDto.getCreateMoneyBox() != null && requestDto.getCreateMoneyBox()) {
@@ -192,7 +195,7 @@ public class BucketListServiceImpl implements BucketListService {
 
         // 삭제되지 않은 해당 회원의 진행중인 버킷리스트 조회 (생성일 기준 내림차순)
         List<BucketList> bucketLists = bucketListRepository.findByMemberAndStatusAndDeletedOrderByCreatedAtDesc(
-                member, BucketListStatus.IN_PROGRESS, false);
+                member, IN_PROGRESS, false);
 
         log.info("진행중인 버킷리스트 목록 조회 완료: 총 {}개", bucketLists.size());
 
@@ -209,7 +212,7 @@ public class BucketListServiceImpl implements BucketListService {
 
         // 삭제되지 않은 해당 회원의 완료된 버킷리스트 조회 (생성일 기준 내림차순)
         List<BucketList> bucketLists = bucketListRepository.findByMemberAndStatusAndDeletedOrderByCreatedAtDesc(
-                member, BucketListStatus.COMPLETED, false);
+                member, COMPLETED, false);
 
         log.info("완료된 버킷리스트 목록 조회 완료: 총 {}개", bucketLists.size());
 
@@ -255,16 +258,16 @@ public class BucketListServiceImpl implements BucketListService {
         // 그룹에 속해있는지 확인
         Group group = currentMember.getGroup();
         if (group == null) {
-            throw new CustomException(ErrorCode.GROUP_NOT_FOUND);
+            throw new CustomException(GROUP_NOT_FOUND);
         }
 
         // 대상 그룹원 조회
         Member targetMember = memberRepository.findById(groupMemberId)
-                .orElseThrow(() -> new CustomException(ErrorCode.MEMBER_NOT_FOUND));
+                .orElseThrow(() -> new CustomException(MEMBER_NOT_FOUND));
 
         // 같은 그룹에 속해있는지 확인
         if (targetMember.getGroup() == null || !targetMember.getGroup().getId().equals(group.getId())) {
-            throw new CustomException(ErrorCode.MEMBER_NOT_FOUND);
+            throw new CustomException(MEMBER_NOT_FOUND);
         }
 
         // 1. 그룹원의 모든 진행중인 버킷리스트 조회 (공개 여부 상관없이)
@@ -273,7 +276,7 @@ public class BucketListServiceImpl implements BucketListService {
         // 2. 본인이 참여자인 비공개 진행중인 버킷리스트 조회
         List<BucketList> participatedBucketLists = bucketListRepository.findByParticipantMemberId(currentMember.getId())
                 .stream()
-                .filter(bl -> bl.getMember().getId().equals(groupMemberId) && bl.getStatus() == BucketListStatus.IN_PROGRESS)
+                .filter(bl -> bl.getMember().getId().equals(groupMemberId) && bl.getStatus() == IN_PROGRESS)
                 .toList();
 
         // 3. 접근 가능한 버킷리스트 필터링 (공개된 것 + 참여자인 것)
@@ -282,7 +285,7 @@ public class BucketListServiceImpl implements BucketListService {
         for (BucketList bucketList : allBucketLists) {
             // 공개된 버킷리스트이거나 본인이 참여자인 경우
             if (bucketList.isPublicFlag() ||
-                participatedBucketLists.stream().anyMatch(p -> p.getId().equals(bucketList.getId()))) {
+                    participatedBucketLists.stream().anyMatch(p -> p.getId().equals(bucketList.getId()))) {
                 accessibleBucketLists.add(bucketList);
             }
         }
@@ -307,16 +310,16 @@ public class BucketListServiceImpl implements BucketListService {
         // 그룹에 속해있는지 확인
         Group group = currentMember.getGroup();
         if (group == null) {
-            throw new CustomException(ErrorCode.GROUP_NOT_FOUND);
+            throw new CustomException(GROUP_NOT_FOUND);
         }
 
         // 대상 그룹원 조회
         Member targetMember = memberRepository.findById(groupMemberId)
-                .orElseThrow(() -> new CustomException(ErrorCode.MEMBER_NOT_FOUND));
+                .orElseThrow(() -> new CustomException(MEMBER_NOT_FOUND));
 
         // 같은 그룹에 속해있는지 확인
         if (targetMember.getGroup() == null || !targetMember.getGroup().getId().equals(group.getId())) {
-            throw new CustomException(ErrorCode.MEMBER_NOT_FOUND);
+            throw new CustomException(MEMBER_NOT_FOUND);
         }
 
         // 1. 그룹원의 모든 완료된 버킷리스트 조회 (공개 여부 상관없이)
@@ -325,7 +328,7 @@ public class BucketListServiceImpl implements BucketListService {
         // 2. 본인이 참여자인 비공개 완료된 버킷리스트 조회
         List<BucketList> participatedBucketLists = bucketListRepository.findByParticipantMemberId(currentMember.getId())
                 .stream()
-                .filter(bl -> bl.getMember().getId().equals(groupMemberId) && bl.getStatus() == BucketListStatus.COMPLETED)
+                .filter(bl -> bl.getMember().getId().equals(groupMemberId) && bl.getStatus() == COMPLETED)
                 .toList();
 
         // 3. 접근 가능한 버킷리스트 필터링 (공개된 것 + 참여자인 것)
@@ -334,7 +337,7 @@ public class BucketListServiceImpl implements BucketListService {
         for (BucketList bucketList : allBucketLists) {
             // 공개된 버킷리스트이거나 본인이 참여자인 경우
             if (bucketList.isPublicFlag() ||
-                participatedBucketLists.stream().anyMatch(p -> p.getId().equals(bucketList.getId()))) {
+                    participatedBucketLists.stream().anyMatch(p -> p.getId().equals(bucketList.getId()))) {
                 accessibleBucketLists.add(bucketList);
             }
         }
@@ -358,16 +361,16 @@ public class BucketListServiceImpl implements BucketListService {
 
         // 삭제되지 않은 버킷리스트 조회 (참여자 정보 포함)
         BucketList bucketList = bucketListRepository.findByIdAndDeletedWithParticipants(bucketListId, false)
-                .orElseThrow(() -> new CustomException(ErrorCode.BUCKET_LIST_NOT_FOUND));
+                .orElseThrow(() -> new CustomException(BUCKET_LIST_NOT_FOUND));
 
         // 1. 본인의 버킷리스트인 경우 접근 거부 (이 API는 남의 버킷리스트 조회용)
         if (bucketList.getMember().getId().equals(member.getId())) {
-            throw new CustomException(ErrorCode.BUCKET_LIST_ACCESS_DENIED);
+            throw new CustomException(BUCKET_LIST_ACCESS_DENIED);
         }
 
         // 접근 권한 검증 (공개 버킷리스트 또는 참여자 권한)
         if (!canAccessOthersBucketList(bucketList, member)) {
-            throw new CustomException(ErrorCode.BUCKET_LIST_ACCESS_DENIED);
+            throw new CustomException(BUCKET_LIST_ACCESS_DENIED);
         }
 
         log.info("그룹원 버킷리스트 조회 완료: ID = {}, 소유자 = {}", bucketListId, bucketList.getMember().getName());
@@ -454,7 +457,7 @@ public class BucketListServiceImpl implements BucketListService {
         Group ownerGroup = owner.getGroup();
 
         if (ownerGroup == null) {
-            throw new CustomException(ErrorCode.GROUP_NOT_FOUND);
+            throw new CustomException(GROUP_NOT_FOUND);
         }
 
         // 기존 참여자들을 모두 비활성화
@@ -473,7 +476,7 @@ public class BucketListServiceImpl implements BucketListService {
 
             // 멤버 조회 및 같은 그룹인지 확인
             Member targetMember = memberRepository.findById(memberId)
-                    .orElseThrow(() -> new CustomException(ErrorCode.MEMBER_NOT_FOUND));
+                    .orElseThrow(() -> new CustomException(MEMBER_NOT_FOUND));
 
             if (targetMember.getGroup() == null || !targetMember.getGroup().getId().equals(ownerGroup.getId())) {
                 log.warn("다른 그룹의 멤버이므로 제외: memberId = {}", memberId);
@@ -530,28 +533,28 @@ public class BucketListServiceImpl implements BucketListService {
         BucketList bucketList = getBucketListWithOwnershipValidation(bucketListId, member.getId());
 
         // 진행중인 버킷리스트
-        if (bucketList.getStatus() == BucketListStatus.IN_PROGRESS) {
+        if (bucketList.getStatus() == IN_PROGRESS) {
             Account moneyBoxAccount = bucketList.getMoneyBoxAccount();
             log.info("연결된 머니박스: {}", moneyBoxAccount.getId());
-            
+
             // 진행중인 버킷리스트인 경우 머니박스의 잔액을 주계좌로 반환
-            if (bucketList.getStatus() == BucketListStatus.IN_PROGRESS) {
+            if (bucketList.getStatus() == IN_PROGRESS) {
                 // 머니박스 → 주계좌 전액 인출
                 BigDecimal withdrawnAmount = transferService.withdrawAllFromMoneyBox(
                         member.getId(),
                         moneyBoxAccount.getId(),
                         bucketListId
                 );
-                
-                log.info("버킷리스트 삭제로 인한 머니박스 잔액 인출 완료: {} → 주계좌, 인출금액: {}", 
+
+                log.info("버킷리스트 삭제로 인한 머니박스 잔액 인출 완료: {} → 주계좌, 인출금액: {}",
                         moneyBoxAccount.getId(), withdrawnAmount);
             }
-            
+
             // 머니박스 계좌 삭제
             moneyBoxAccount.setDeleted(true);
             accountService.save(moneyBoxAccount);
             log.info("머니박스 계좌 삭제 완료: {}", moneyBoxAccount.getId());
-            
+
             // 관련 자동이체 스케줄 모두 삭제 및 비활성화
             autoTransferScheduleService.deleteAllSchedulesForMoneyBox(moneyBoxAccount);
             log.info("자동이체 스케줄 삭제 완료");
@@ -570,25 +573,25 @@ public class BucketListServiceImpl implements BucketListService {
     private BigDecimal calculateInterestRate(Integer targetMonth, Account mainAccount) {
         // 기본 이자율 계산
         BigDecimal baseRate = switch (targetMonth) {
-            case 3  -> BigDecimal.valueOf(0.010); // 1.0%
-            case 6  -> BigDecimal.valueOf(0.015); // 1.5%
+            case 3 -> BigDecimal.valueOf(0.010); // 1.0%
+            case 6 -> BigDecimal.valueOf(0.015); // 1.5%
             case 12 -> BigDecimal.valueOf(0.025); // 2.5%
             case 24 -> BigDecimal.valueOf(0.032); // 3.2%
             default -> BigDecimal.valueOf(0.010); // 기본값: 3개월
         };
-        
+
         // 연금통장 우대 이자율 적용 (+1.0%)
         if (mainAccount.getName().contains("연금")) {
             BigDecimal originalRate = baseRate;
             baseRate = baseRate.add(BigDecimal.valueOf(0.010));
-            log.info("연금통장 우대이자 적용 - 기간: {}개월, 기본: {}% + 우대: 1.0% = 최종: {}%", 
-                    targetMonth, originalRate.multiply(BigDecimal.valueOf(100)), 
+            log.info("연금통장 우대이자 적용 - 기간: {}개월, 기본: {}% + 우대: 1.0% = 최종: {}%",
+                    targetMonth, originalRate.multiply(BigDecimal.valueOf(100)),
                     baseRate.multiply(BigDecimal.valueOf(100)));
         } else {
-            log.info("이자율 계산 완료 - 기간: {}개월, 이자율: {}%", 
+            log.info("이자율 계산 완료 - 기간: {}개월, 이자율: {}%",
                     targetMonth, baseRate.multiply(BigDecimal.valueOf(100)));
         }
-        
+
         return baseRate;
     }
 
@@ -601,21 +604,21 @@ public class BucketListServiceImpl implements BucketListService {
         BucketList bucketList = getBucketListWithOwnershipValidation(bucketListId, member.getId());
 
         // 이미 완료된 버킷리스트인지 확인
-        if (bucketList.getStatus() == BucketListStatus.COMPLETED) {
-            throw new CustomException(ErrorCode.BUCKET_LIST_ALREADY_COMPLETED);
+        if (bucketList.getStatus() == COMPLETED) {
+            throw new CustomException(BUCKET_LIST_ALREADY_COMPLETED);
         }
 
         // 오늘 날짜가 targetDate 이전이면 아직 달성 불가
         if (LocalDate.now().isBefore(bucketList.getTargetDate())) {
-            throw new CustomException(ErrorCode.BUCKET_LIST_NOT_YET_AVAILABLE);
+            throw new CustomException(BUCKET_LIST_NOT_YET_AVAILABLE);
         }
 
         // 주계좌 조회
         Account mainAccount = accountService.findMainAccountByMember(member);
-        
+
         // 연결된 머니박스가 있는지 확인
         Account moneyBoxAccount = bucketList.getMoneyBoxAccount();
-        
+
         if (moneyBoxAccount != null) {
             // 1. 머니박스 → 주계좌로 원금 인출
             BigDecimal withdrawnAmount = transferService.withdrawAllFromMoneyBox(
@@ -623,9 +626,9 @@ public class BucketListServiceImpl implements BucketListService {
                     moneyBoxAccount.getId(),
                     bucketListId
             );
-            log.info("목표 달성 원금 인출 완료: 머니박스 {} → 주계좌, 인출금액: {}", 
+            log.info("목표 달성 원금 인출 완료: 머니박스 {} → 주계좌, 인출금액: {}",
                     moneyBoxAccount.getId(), withdrawnAmount);
-            
+
             // 2. 이자 계산 및 지급 (목표금액 한도 내에서만)
             BigDecimal interestRate = calculateInterestRate(bucketList.getTargetMonth(), mainAccount);
 
@@ -641,12 +644,12 @@ public class BucketListServiceImpl implements BucketListService {
             if (interest.compareTo(BigDecimal.ZERO) > 0) {
                 transferService.payInterest(member.getId(), interest, bucketListId);
             }
-            
+
             // 3. 머니박스 계좌 삭제
             moneyBoxAccount.setDeleted(true);
             accountService.save(moneyBoxAccount);
             log.info("머니박스 계좌 삭제 완료: {}", moneyBoxAccount.getId());
-            
+
             // 4. 관련 자동이체 스케줄 모두 삭제 및 비활성화
             autoTransferScheduleService.deleteAllSchedulesForMoneyBox(moneyBoxAccount);
             log.info("자동이체 스케줄 삭제 완료");
@@ -654,7 +657,7 @@ public class BucketListServiceImpl implements BucketListService {
 
         // 5. 버킷리스트 상태를 완료로 변경
         BucketListStatus previousStatus = bucketList.getStatus();
-        bucketList.setStatus(BucketListStatus.COMPLETED);
+        bucketList.setStatus(COMPLETED);
         BucketList savedBucketList = bucketListRepository.save(bucketList);
 
         // 6. 쿠폰 발행
